v1.0!!!!
Hooray! Don't Burn Your Breakfast is ready for release!! (Probably). There's been a lot of polish since it was "content complete", and it's in a really good state now! I'm sure I'll need to tweak it slightly based on player feedback, but hopefully it's only minor changes from here on out!